requirements.txt
Flask==0.10.1
Flask-MySQLdb==0.2.0
Flask-Session==0.3.0
gunicorn==18.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.7
Werkzeug==0.9.4
